Fix reading environment variables in deployment.toml
The deployment.toml file is parsed using the net.consensys.cava.toml.Toml parser which doesn't consider $env{} as environment variable.
But there is already the toml file is parsed using the ConfigParser which supports reading environemnt variables. This results is not used in all places.
So, the result from the ConfigParser is used in all the places in this fix.
It fixes  https://github.com/wso2/micro-integrator/issues/2964